# swagger_full.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Flash API",
    "version": "1.0.0",
    "description": "API completa para gerenciamento de múltiplas sessões WhatsApp",
    "contact": {
      "name": "API Support",
      "email": "quickgestor@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de Beta"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey",
        "description": "API Key para autenticação"
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/config/config": {
      "put": {
        "summary": "Atualizar configurações da sessão",
        "tags": [
          "Configurações"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ignoreGroups": {
                    "type": "boolean",
                    "description": "Ignorar mensagens de grupos",
                    "default": false
                  },
                  "autoRead": {
                    "type": "boolean",
                    "description": "Marcar mensagens como lidas automaticamente",
                    "default": false
                  },
                  "rejectCalls": {
                    "type": "boolean",
                    "description": "Rejeitar chamadas automaticamente",
                    "default": true
                  },
                  "msg_rejectCalls": {
                    "type": "string",
                    "description": "Resposta De regeitar msg",
                    "default": "Não atendemos ligações"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configurações atualizadas com sucesso"
          }
        }
      }
    },
...
    "/session/create_sessao": {
      "post": {
        "summary": "Criar Sessao",
        "tags": [
          "Sessões"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "sessionId"
                ],
                "properties": {
                  "nome_sessao": {
                    "type": "string",
                    "description": "Nome da sessão (obrigatório, mínimo 5 caracteres)"
                  },
                  "numero": {
                    "type": "string",
                    "description": "Numero para contato"
                  },
                  "criar_sessao": {
                    "type": "boolean",
                    "description": "Se true, Criar a conexão com whatsapp",
                    "default": true
                  },
                  "gerar_qrcode": {
                    "type": "boolean",
                    "description": "Se true, Gerar qrcode Base64",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sessão criada com sucesso"
          }
        }
      }
    },
    "/session/conectar_sessao": {
      "put": {
        "summary": "Conectar a uma sessão WhatsApp existente",
        "tags": [
          "Sessões"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Você recebe um base64 com qrcode"
          }
        }
      }
    },
    "/session/restart": {
      "put": {
        "summary": "Reniciar uma sessão conectada",
        "tags": [
          "Sessões"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Você recebe uma mensagem de sucesso"
          }
        }
      }
    },
    "/session/status/": {
      "get": {
        "summary": "Obter status da sessão",
        "tags": [
          "Sessões"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Status obtido com sucesso"
          }
        }
      }
    },
    "/session/list": {
      "get": {
        "summary": "Listar todas as sessões",
        "tags": [
          "Sessões"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de sessões"
          }
        }
      }
    },
    "/session/reconnect": {
      "post": {
        "summary": "Forçar reconexão da sessão",
        "tags": [
          "Sessões"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Reconexão iniciada com sucesso"
          }
        }
      }
    },
...
    "/session/delete/{sessionId}": {
      "delete": {
        "summary": "Deletar sessão",
        "tags": [
          "Sessões"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sessão deletada com sucesso"
          }
        }
      }
    },
...
    "/system/status": {
      "get": {
        "summary": "Status geral do sistema",
        "tags": [
          "Sistema"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Status do sistema"
          }
        }
      }
    },
    "/system/config": {
      "get": {
        "summary": "Configurações do sistema",
        "tags": [
          "Sistema"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Configurações do sistema"
          }
        }
      }
    },
    "/system/cleanup": {
      "post": {
        "summary": "Limpar sessões desconectadas",
        "tags": [
          "Sistema"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Limpeza realizada com sucesso"
          }
        }
      }
    }
  }
}